+++ src/Utilities/Enum.h
@@ -32,6 +32,7 @@
 
 #pragma once
 
+#include <GeneralDefinitions.h>
 #include <Phobos.h>
 #include <GeneralDefinitions.h>
 
@@ -194,6 +195,14 @@ enum class TargetZoneScanType
 	Any = 1,
 	InRange = 2
 };
+
+enum class AttachmentYSortPosition
+{
+	Default = 0,
+	UnderParent = 1,
+	OverParent = 2
+};
+
 
 enum class ChronoSparkleDisplayPosition : unsigned char
 {