+++ docs/Whats-New.md
@@ -266,6 +266,15 @@ You can use the migration utility (can be found on [Phobos supplementaries repo]
 
 ## Changelog
 
+### 0.4
+
+<details>
+  <summary>Click to show</summary>
+
+New:
+- Techno Attachment logic (by Kerbiter)
+</details>
+
 ### Version TBD (develop branch nightly builds)
 
 <details open>